cyber-forensics-toolkit/requirements.txt
PyQt5==5.15.9
PyPDF2
python-docx
pillow
pytsk3
